Changed fabsf to fabs. Original: 302cd3392c97e7607c8a0914ddf65cff4a163db8

+++ lal/packages/tools/src/SnglInspiralUtils.c
@@ -202,8 +202,8 @@ LALCompareSnglInspiral (
 
   if ( labs(ta-tb) < params->dtime )
   {
-    dm1 = fabsf( aPtr->mass1 - bPtr->mass1 );
-    dm2 = fabsf( aPtr->mass2 - bPtr->mass2 );
+    dm1 = fabs( aPtr->mass1 - bPtr->mass1 );
+    dm2 = fabs( aPtr->mass2 - bPtr->mass2 );
 
     if ( dm1 <= params->dm && dm2 <= params->dm )
     {
